Separate App into components
*Add links to the bar.
*Separate the timeline from App into its own component.
*Restructure files and folders.
*Create Admin component.
