+++ src/Pages/Admin/Admin.js
@@ -1,0 +1,26 @@
+import React from 'react'
+
+class Admin extends React.Component {
+    constructor(props) {
+        super(props)
+    }
+
+    render() {
+        console.log(this.props)
+        if (!this.props.admin) {
+            return (
+                <div>
+                    Du är inte admin
+                </div>
+            )
+        } else {
+            return (
+                <div>
+                    Du är admin
+                </div>
+            )
+        }
+    }
+}
+
+export default Admin
